requirement.txt
1-yfinance
2-streamlit
